intelligence_worker/requirements.txt
psycopg2-binary>=2.9
tenacity>=8.2
redis>=4.6
